backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib==1.7.4
bcrypt==3.2.0
psycopg2-binary
python-multipart

itsdangerous
